Double Six Dice Game
The first player rolls a pair of dice, and keeps rolling the dice again and again, until they roll a double six.
It is then the turn of the the second player. They too will roll the dice and keep doing so until they roll a double six.
The winner of the game is the user who rolled a double six in the smaller number of attempts.
The game ends on a draw if both players used the same number of attempts to roll a double six.